refactor: standardize conventions for all base components
- move all exports to index.ts
- move all types to types.ts
- prefer "props" over "attributes" for props spreading
- prefer React.ComponentPropsWithoutRef for base HTML interfaces
