+++ packages/react/src/components/BaseButton/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseButton';
+export * from './types';
