+++ packages/react/src/components/BaseDetails/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseDetails';
+export * from './types';
