+++ packages/react/src/components/BaseDetails/types.ts
@@ -1,0 +1,1 @@
+export type BaseDetailsProps = React.ComponentPropsWithoutRef<'details'>;
