+++ packages/react/src/components/BaseDialog/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseDialog';
+export * from './types';
