+++ packages/react/src/components/BaseDialog/types.ts
@@ -1,0 +1,3 @@
+export interface BaseDialogProps extends React.ComponentPropsWithoutRef<'div'> {
+	modal?: boolean;
+}
