+++ packages/react/src/components/BaseInput/BaseInput.tsx
@@ -4,52 +4,57 @@ import {
 	useLayoutEffect,
 	useValidation,
 } from '../../utilities';
-import { BaseTextAreaProps } from './types';
+import { BaseInputProps } from './types';
 
-const defaultProps: BaseTextAreaProps = {
+const defaultProps: BaseInputProps = {
 	validateOnDOMChange: true,
 };
 
 /**
- * A base `<textarea>` component. Adds a callback for the DOM's `change` event
+ * A base `<input>` component. Adds a callback for the DOM's `change` event
  * (`onDOMChange`), which does not exist in React.
  */
-export const BaseTextArea = React.forwardRef<HTMLTextAreaElement, BaseTextAreaProps>(({
-	multiline = false,
-	autoSize = false,
+export const BaseInput = React.forwardRef<HTMLInputElement, BaseInputProps>(({
 	errors: errorsProp,
 	validateOnChange,
 	validateOnDOMChange = defaultProps.validateOnDOMChange,
 	validators,
-	// pull out maxLength because it prevents user textarea past the given
+	// pull out maxLength because it prevents user input past the given
 	// length, which is an anti-pattern according to our usage guidelines.
 	maxLength,
 	maxLengthRestrictsInput = false,
 	onInput,
 	onDOMChange,
 	onValidate,
-	...attributes
-}: BaseTextAreaProps, ref): React.ReactElement => {
-	const [textarea, setTextarea] = useForwardedRef(ref);
+	...props
+}: BaseInputProps, ref): React.ReactElement => {
+	const [input, setInput] = useForwardedRef(ref);
 	const [errors, setErrors] = React.useState(errorsProp);
-	const lines = Number(multiline) > 0 ? Number(multiline) : 1;
 
 	// treat the prop version of `errors` as the source of truth
 	React.useEffect(() => setErrors(errorsProp), [errorsProp]);
 
 	const validator = useValidation(validators);
 	const validate = React.useCallback(({
+		max,
 		maxLength: elMaxLength,
+		min,
 		minLength,
+		pattern,
 		required,
+		step,
 		type,
 		value,
 		validity,
-	}: HTMLTextAreaElement) => {
+	}: HTMLInputElement) => {
 		const errs = validator({
+			max,
 			maxLength: maxLength || elMaxLength,
+			min,
 			minLength,
+			pattern,
 			required,
+			step,
 			type,
 			value,
 			validity,
@@ -58,64 +63,51 @@ export const BaseTextArea = React.forwardRef<HTMLTextAreaElement, BaseTextAreaPr
 		if (!errorsProp) setErrors(errs);
 	}, [validator, onValidate, errorsProp, maxLength]);
 
-	// Handles auto sizing with multiline component
-	React.useEffect(() => {
-		if (autoSize && textarea) textarea.style.height = 'auto';
-	}, [autoSize, textarea]);
-
-	const onResize = () => {
-		if (textarea) {
-			textarea.style.height = 'auto';
-			textarea.style.height = `${textarea.scrollHeight}px`;
-		}
-	};
-
 	/**
 	 * Unlike `onChange`, `onInput` will trigger even when the user enters a bad
 	 * value, such as entering a letter in a `type="number"` field, so run
 	 * validation here to catch the `ValidityState.badInput` errors.
 	 */
-	const inputHandler = (e: React.FormEvent<HTMLTextAreaElement>): void => {
+	const inputHandler = (e: React.FormEvent<HTMLInputElement>): void => {
 		if (onInput) onInput(e);
-		if (autoSize) onResize();
+
 		if (validateOnChange) validate(e.currentTarget);
 	};
 
 	const domChangeHandler = React.useCallback((e: Event): void => {
 		if (onDOMChange) onDOMChange(e);
-		if (validateOnDOMChange) validate(e.target as HTMLTextAreaElement);
+		if (validateOnDOMChange) validate(e.target as HTMLInputElement);
 	}, [onDOMChange, validateOnDOMChange, validate]);
 
 	// Reflect errors on the DOM's constraint validation API. This ensures that
 	// browser tooltip text always matches the custom errors.
 	React.useEffect(() => {
-		if (textarea && textarea.willValidate) {
+		if (input && input.willValidate) {
 			const errString = (!errors || !errors.length) ? '' : errors.join('\n');
-			textarea.setCustomValidity(errString);
+			input.setCustomValidity(errString);
 		}
-	}, [textarea, errors]);
+	}, [input, errors]);
 
 	// Polyfill the DOM `change` listener
 	useLayoutEffect(() => {
-		if (textarea && domChangeHandler) {
-			textarea.addEventListener('change', domChangeHandler);
+		if (input && domChangeHandler) {
+			input.addEventListener('change', domChangeHandler);
 		}
 		return (): void => {
-			if (textarea && domChangeHandler) {
-				textarea.removeEventListener('change', domChangeHandler);
+			if (input && domChangeHandler) {
+				input.removeEventListener('change', domChangeHandler);
 			}
 		};
-	}, [textarea, domChangeHandler]);
+	}, [input, domChangeHandler]);
 
 	return (
-		<textarea
-			ref={setTextarea}
-			rows={lines}
+		<input
+			ref={setInput}
 			onInput={inputHandler}
 			maxLength={(maxLengthRestrictsInput) ? maxLength : undefined}
-			{...attributes}
+			{...props}
 		/>
 	);
 });
 
-BaseTextArea.defaultProps = defaultProps;
+BaseInput.defaultProps = defaultProps;
