+++ packages/react/src/components/BaseInput/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseInput';
+export * from './types';
