+++ packages/react/src/components/BaseProgress/BaseProgress.tsx
@@ -1,15 +1,5 @@
 import React from 'react';
-
-export interface BaseProgressProps extends React.HTMLAttributes<HTMLDivElement> {
-	/** Indicates how much of the task has been completed. */
-	progress?: number;
-	/**
-	 * Describes how much work is required to complete the progress. If defined,
-	 * it must be greater than `0` and be a valid floating point number.
-	 * The default is `1`.
-	 */
-	max?: number;
-}
+import { BaseProgressProps } from './types';
 
 /**
  * A minimal progress indicator.
@@ -29,8 +19,7 @@ export const BaseProgress = React.forwardRef<HTMLDivElement, BaseProgressProps>(
 		role="progressbar"
 		aria-valuenow={progress}
 		aria-valuemax={max}
-		// min is always 0 for <progress> so we're mirroring that here
-		aria-valuemin={0}
+		aria-valuemin={0}	// min is always 0 for <progress> so we're mirroring that here
 		ref={ref}
 		{...props}
 	>
