+++ packages/react/src/components/BaseProgress/index.ts
@@ -1,4 +1,4 @@
 export * from './BaseProgress';
 export * from './BaseProgressBar';
 export * from './BaseProgressSpinner';
-export * from './CoreProgressProps';
+export * from './types';
