+++ packages/react/src/components/BaseSVG/BaseSVG.tsx
@@ -1,20 +1,16 @@
 import React from 'react';
-
-export interface BaseSVGProps extends React.SVGAttributes<SVGSVGElement> {
-	/** A URL or reference to the source of this SVG. */
-	source?: string;
-}
+import { BaseSVGProps } from './types';
 
 export const BaseSVG = React.forwardRef<SVGSVGElement, BaseSVGProps>(({
 	source,
 	children,
-	...attributes
+	...props
 }: BaseSVGProps, ref) => (
 	<svg
 		ref={ref}
 		xmlns="http://www.w3.org/2000/svg"
 		data-source={source}
-		{...attributes}
+		{...props}
 	>
 		{ children }
 	</svg>
