+++ packages/react/src/components/BaseSVG/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseSVG';
+export * from './types';
