+++ packages/react/src/components/BaseSVG/types.ts
@@ -1,0 +1,4 @@
+export interface BaseSVGProps extends React.ComponentPropsWithoutRef<'svg'> {
+	/** A URL or reference to the source of this SVG. */
+	source?: string;
+}
