+++ packages/react/src/components/BaseSummary/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseSummary';
+export * from './types';
