+++ packages/react/src/components/BaseTextArea/index.ts
@@ -1,0 +1,2 @@
+export * from './BaseTextArea';
+export * from './types';
